+++ src/tests/unittests/algorithms/test_functions.py
@@ -1,13 +1,22 @@
-""" Mathematical functions and models."""
-
 import unittest
 import matplotlib.pyplot as plt
 import numpy as np
+import qtt
 from qtt.algorithms.functions import gaussian, fit_gaussian, fit_double_gaussian, double_gaussian, exp_function, \
-    fit_gauss_ramsey, gauss_ramsey, cost_exp_decay, logistic, linear_function, Fermi, fit_exp_decay, _estimate_exp_decay_initial_parameters
+    fit_gauss_ramsey, gauss_ramsey, cost_exp_decay, logistic, linear_function, Fermi, fit_exp_decay, \
+    _estimate_exp_decay_initial_parameters, plot_gauss_ramsey_fit
 
 
 class TestFunctions(unittest.TestCase):
+
+    def test_estimate_dominant_frequency(self):
+        y_data = np.array([0.122, 0.2  , 0.308, 0.474, 0.534, 0.618, 0.564, 0.436, 0.318,
+               0.158, 0.13 , 0.158, 0.336, 0.434, 0.51 , 0.59 , 0.592, 0.418,
+               0.286, 0.164, 0.156, 0.186, 0.25 , 0.362, 0.524])
+        sample_rate = 47368421
+        estimated_frequency = qtt.algorithms.functions.estimate_dominant_frequency(y_data, sample_rate = sample_rate, fig=1)
+        self.assertAlmostEqual(estimated_frequency, 5684210, places=-1)
+        plt.close(1)
 
     @staticmethod
     def test_fit_exp_decay():
@@ -95,24 +104,13 @@ class TestFunctions(unittest.TestCase):
                            0.3061, 0.3161, 0.3976, 0.4246, 0.398, 0.3757, 0.3615, 0.3723,
                            0.3803, 0.3873, 0.3873, 0.3561, 0.37, 0.3819, 0.3834, 0.3838,
                            0.37, 0.383, 0.3573, 0.3869, 0.3838, 0.3792, 0.3757, 0.3815])
-        x_data = np.array([i * 1.6 / 40 for i in range(40)])
+        x_data = 1e-6*np.array([i * 1.6 / 40 for i in range(40)])
 
-        par_fit_test, _ = fit_gauss_ramsey(x_data * 1e-6, y_data)
+        par_fit_test, _ = fit_gauss_ramsey(x_data, y_data)
 
         self.assertTrue(np.abs(np.abs(par_fit_test[0]) - 0.21) < 0.1)
-        self.assertTrue(np.abs(par_fit_test[-2] - 1.88) < 0.1)
+        self.assertTrue(np.abs(par_fit_test[-2] - (-1.255)) < 0.1)
         self.assertTrue(np.abs(par_fit_test[-1] - 0.38) < 0.1)
-
-        test_x = np.linspace(0, x_data.max() * 1e-6, 200)
-        test_y = gauss_ramsey(test_x, par_fit_test)
-
-        if fig is not None:
-            plt.figure(fig)
-            plt.clf()
-            plt.plot(x_data, y_data, 'o', label='input data')
-            plt.plot(test_x * 1e6, test_y, label='fitted curve')
-            plt.legend(numpoints=1)
-            plt.close('all')
 
     def test_logistic_and_linear_function(self):
         x_data = np.arange(-10, 10, 0.1)